Split migrate baseline black-box test
